+++ src/content/config.ts
@@ -31,6 +31,7 @@ const changelogs = defineCollection({
       .object({
         path: z.string(),
         absoluteFilePath: z.string(),
+        astroContentFilePath: z.string(),
         filePath: z.string(),
         publicPath: z.string(),
         type: z.string(),
@@ -60,10 +61,12 @@ const baseSchema = z.object({
   // Used by eventcatalog
   versions: z.array(z.string()).optional(),
   latestVersion: z.string().optional(),
+  pathToFile: z.string().optional(),
   catalog: z
     .object({
       path: z.string(),
       filePath: z.string(),
+      astroContentFilePath: z.string(),
       publicPath: z.string(),
       type: z.string(),
     })